Conflicts:
	src/game/AuctionHouseBot.cpp
	src/game/BattleGround.cpp
	src/game/Language.h
	src/game/Level2.cpp
	src/game/Player.cpp
	src/game/Spell.cpp
	src/game/SpellAuras.cpp
	src/game/SpellEffects.cpp
	src/game/SpellMgr.cpp
	src/game/Unit.cpp
	src/game/UnitAuraProcHandler.cpp
	src/mangosd/mangosd.conf.dist.in

+++ src/game/LFGMgr.cpp
@@ -2062,7 +2062,7 @@ bool LFGMgr::CheckRoles(LFGRolesMap* rolesMap)
 //    if (sWorld.getConfig(CONFIG_BOOL_LFG_DEBUG_ENABLE))
 //        return true;
 
-    if ((healers + tanks + dps) > (MAX_GROUP_SIZE - rolesMap->size()))
+    if ((healers + tanks + dps) > uint8(MAX_GROUP_SIZE - rolesMap->size()))
         return false;
 
     return true;
